% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_bandit_values.R
\name{appendBanditValues}
\alias{appendBanditValues}
\title{Append Input from n Armed Bandit Task to Task Control List}
\usage{
appendBanditValues(ctrlVals, distList, input, gameData)
}
\arguments{
\item{ctrlVals}{list of reactive values. Controls the flow through the
Experiment. Can be created with \code{\link{createCtrlList}}.}

\item{distList}{list. At least contains the number of trials "nTrials"
(usually the one used to create the distributions with
\code{\link{createBanditList}}).}

\item{input}{The input object from a shiny app.}

\item{gameData}{list of reactive values. Specifically for the bandit task.
Can be created with \code{\link{createTaskCtrlList}}.}
}
\value{
The changed gameData list of reactive values (note: does not have
 to be assigned when function is called).
}
\description{
After a game in the bandit task is ended the data will be sent from the browser
to the current R session. This functions appends the sent data to task control
list (gameData) which has to be a reactive value (see
\code{\link[shiny]{reactiveValues}} or
\url{https://shiny.rstudio.com/reference/shiny/latest/reactiveValues.html}).
}
