% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_dd_values.R
\name{appendDdValues}
\alias{appendDdValues}
\title{Append Input from Delay Discounting Task to Task Control List}
\usage{
appendDdValues(ctrlVals, input, gameData, container,
  withPracticeTrial = TRUE, afterPracticePage = "postPractice",
  afterTrialPage = "endGame", afterLastTrialPage = "lastEndGame")
}
\arguments{
\item{ctrlVals}{list of reactive values. Controls the flow through the
Experiment. Can be created with \code{\link{createCtrlList}}.}

\item{input}{The input object from a shiny app.}

\item{gameData}{list of reactive values. Specifically for the bandit task.
Can be created with \code{\link{createTaskCtrlList}}.}

\item{container}{list. Containing at least an object named "trial.order".
Usually this list is created with \code{\link{createDdList}}.}

\item{withPracticeTrial}{logical. If TRUE it is assumed that a practice trial
was included. The page after that trial will be afterPracticePage.}

\item{afterPracticePage}{string. The name of the page to be shown
after the practice trial. Only needs to be specified if withPracticeTrial
is set to TRUE.}

\item{afterTrialPage}{string. The name of the page to be shown
after the trials up to the second to last trial.}

\item{afterLastTrialPage}{string. The name of the page to be shown
after the last trial..}
}
\value{
The changed gameData list of reactive values and change current page
value (note: does not have to be assigned when function is called).
}
\description{
After a trial in the delay discouting task is ended the data is appended to
the task control list (gameData) which has to be a reactive value (see
\code{\link[shiny]{reactiveValues}} or
\url{https://shiny.rstudio.com/reference/shiny/latest/reactiveValues.html}).
}
