% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_dfd_values.R
\name{appendDfdValues}
\alias{appendDfdValues}
\title{Append Input from Decisions from Description Task to Task Control List}
\usage{
appendDfdValues(ctrlVals, input, gameData, container,
  withPracticeGamble = TRUE, afterPracticePage = "postPractice",
  afterGamblePage = "endGame", afterLastGamblePage = "lastEndGame")
}
\arguments{
\item{ctrlVals}{list of reactive values. Controls the flow through the
Experiment. Can be created with \code{\link{createCtrlList}}.}

\item{input}{The input object from a shiny app.}

\item{gameData}{list of reactive values. Specifically for the bandit task.
Can be created with \code{\link{createTaskCtrlList}}.}

\item{container}{list. Containing at least two objects named "trial.order"
and "gamle.order". Usually this list is created with \code{\link{createDfdList}}.}

\item{withPracticeGamble}{logical. If TRUE it is assumed that a practice
gamble was included. The page after that trial will be afterPracticePage.}

\item{afterPracticePage}{string. The name of the page to be shown
after the practice trial. Only needs to be specified if withPracticeGamble
is set to TRUE.}

\item{afterGamblePage}{string. The name of the page to be shown
after the trials up to the second to last trial.}

\item{afterLastGamblePage}{string. The name of the page to be shown
after the last trial.}
}
\value{
changed gameData list of reactive values and change current page
value (note: does not have to be assigned when function is called).
}
\description{
After a gamble in the decisions from describtion task is ended the data is
appended to the task control list (gameData) which has to be a reactive value
(see \code{\link[shiny]{reactiveValues}} or
\url{https://shiny.rstudio.com/reference/shiny/latest/reactiveValues.html}).
}
