% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_page_variable.R
\name{changePageVariable}
\alias{changePageVariable}
\title{Change a Variable from a Page List that}
\usage{
changePageVariable(pageList, variable, oldLabel, newLabel)
}
\arguments{
\item{pageList}{list. Contains information to build a shiny page. Usually
created with \code{\link{createPageList}}.}

\item{variable}{string. The name of the variable to be changed.}

\item{oldLabel}{string. Must match the current value of the pageList that
should be changed.}

\item{newLabel}{An object that will replace the oldLabel in the pageList.}
}
\value{
The given pageList with the updated value.
}
\description{
Changes a variable from a page list has to be changed e.g. because a code was
generated in the current r session.
}
