% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ID.R
\name{checkId}
\alias{checkId}
\title{Check if a User is Allowed to Participate Given his Id}
\usage{
checkId(id, cntrlVals, location = "local", fileSep = ",",
  fileHeader = TRUE, dropDir = NULL, fileName = NULL,
  droptoken = NULL, idsVec = NULL, notAllowedId = "not allowed")
}
\arguments{
\item{id}{string or numeric. The user id to check.}

\item{cntrlVals}{list of reactive values.Controls the flow through the
Experiment. Can be created with \code{\link{createCtrlList}}.}

\item{location}{string. Where to get the reference database from. Must be one
of "local", "dropbox" or "vector". "local" will load a txt or rds file.
"dropbox" will load a csv file from the indicated dropbox account and "vector"
assumes that you gave a vector of ids as input to idsVec.}

\item{fileSep}{string. Passed to the \code{\link{read.table}} sep argument.
Must match the separater used in the database file.}

\item{fileHeader}{logical. Passed to the read.table header argument.}

\item{dropDir}{string. The directory in the dropbox from which to read in the
database. Must only be specified if location is set to "dropbox".}

\item{fileName}{string. The database file-name, including path. Must be
specified if location is one of "dropbox" or "local".}

\item{droptoken}{string. The name of the rds file containing the access tokens
of the dropbox account from which the data should be loaded. If file is
outside the app directory full directory must be included.}

\item{idsVec}{string. Vector of ids that are no longer allowed to participate.
Must only be specified if location is set to "vector".}

\item{notAllowedId}{string. The name (default is "not allowed") of the page
that should be displayed in case the given id is occurring in the database.}
}
\value{
If id is also contained in database cntrlVals$page will be changed to
 notAllowedId. Does not have to be reassigned since it's a reactive value.
}
\description{
Check if a User is Allowed to Participate Given his Id
}
