% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bandit_ctrl_list.R
\name{createBanditCtrlList}
\alias{createBanditCtrlList}
\title{Create a List of Reactive Values for Bandit Task Control}
\usage{
createBanditCtrlList(oVar = NULL, oVarVals = NULL)
}
\arguments{
\item{oVar}{string. Vector with names of additional variables you want to
add to the control list.}

\item{oVarVals}{string or numeric. Vector with values that should be appended
to the list of reactive values on the positions specified in oVar.}
}
\value{
A list of reactive values.
}
\description{
Create a list of reactive values (see \code{\link[shiny]{reactiveValues}} or
\url{https://shiny.rstudio.com/reference/shiny/latest/reactiveValues.html})
to later store the task values in. Additional variables can be specified.
}
