% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bart_page.R
\name{createBartPage}
\alias{createBartPage}
\title{Create an HTML Page to Display and Play the BART}
\usage{
createBartPage(ctrlVals, session, balloonList, RDrawBoundary = TRUE)
}
\arguments{
\item{ctrlVals}{list of reactive values. Controls the flow through the
Experiment. Can be created with \code{\link{createCtrlList}}.}

\item{session}{The session object from a shiny app.}

\item{balloonList}{list. Contains information about the balloons, such as
the pop values. Can be created with \code{\link{createBartList}}.}

\item{RDrawBoundary}{logical. If TRUE (default) an outer boundary is drawn
to indicate the maximum possible pump value (not the actual pop
value of that respective balloon!).}
}
\value{
An html page displayed in the shiny app.
}
\description{
Create an HTML page by using the shiny package and send balloon information
to javascript. To get the ctrlVals list call \code{\link{createCtrlList}} and
to for the balloonList call \code{\link{createBartList}}. The function only
works from within a shiny app (i.e. html code will also be produced when
this function is called from outside a shiny app but nothing will be
displayed).
}
