% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_not_allowed_page.R
\name{createNotAllowedPage}
\alias{createNotAllowedPage}
\title{Create HTML Page for Users not Allowed to Participate}
\usage{
createNotAllowedPage(inputList, globId, idLabel = "WorkerID",
  idVar = "workerid")
}
\arguments{
\item{inputList}{The input object from a shiny app.}

\item{globId}{string. The global id given to the page on which the participant
id was entered.}

\item{idLabel}{string. The id label you want the participant to see,
e.g. if this is set to "WorkerId" the participant will (among other things)
see "You entered the WorkerID ...".}

\item{idVar}{string. The id label specified for the \code{\link[shiny]{textInput}}
in the file given to \code{\link{createPageList}}.}
}
\value{
An html page that can be displayed by shiny.
}
\description{
If you call \code{\link{checkId}} to make sure a user is acutally allowed to
participate but he is already in your database and therefore not allowed
to participate a second time you can then send him to this page. It does not
include a continue button, i.e. you can't navigate any further through the
experiment.
}
