% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_page.R
\name{createPage}
\alias{createPage}
\title{Create HTML Logic for a Page Prepared with \code{\link{createPageList}}}
\usage{
createPage(pageList, pageNumber, globId, ctrlVals, continueButton = TRUE)
}
\arguments{
\item{pageList}{list. Either returned by \code{\link{createPageList}} or of
the same structure.}

\item{pageNumber}{integer. The current page number of this page id.
Usually the one returned by \code{\link{createCtrlList}} where
the page id was specified in the globIds argument.}

\item{globId}{string. The global Id of the current page or questionnaire.}

\item{ctrlVals}{list of reactive values. Controls the flow through the
Experiment. Can be created with \code{\link{createCtrlList}}.}

\item{continueButton}{logical. If TRUE (default) a continue button is added
at the bottom of the page.}
}
\value{
An html page that can be displayed by the shiny app.
}
\description{
Create a html page from the list given to pageList.
}
