% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_page_list.R
\name{createPageList}
\alias{createPageList}
\title{Read in Information on What to Display on a Page and Prepare it for Later Use.}
\usage{
createPageList(defaulttxt = TRUE, location = "local",
  fileName = "Example", randomize = TRUE, globId = fileName,
  droptoken = "droptoken.rds")
}
\arguments{
\item{defaulttxt}{logical. If TRUE (default) a default option is used.}

\item{location}{string. The file location. Valid inputs are
"local" or "dropbox". Not needed if defaulttxt is TRUE.}

\item{fileName}{string. The name of the file to read in. If
defaulttxt is TRUE this has to be one of the options specified in details.
Else it has to be the file name inclusive full path in either dropbox or the
local system.}

\item{randomize}{logical. If TRUE (default) the page variable of
the items that have a 1 in the randomize variable should be randomly ordered.}

\item{globId}{string. Vector containing the names of the page or questionnaire
variables.}

\item{droptoken}{string. The name of the file in which the access tokens for
dropbox are stored. If this file is not in the app directory the full path
must be specified. Must be an rds file.}
}
\value{
A list containing many different parameters necessary to later create
 the page with \code{\link{createPage}}. See example for details.
}
\description{
Reads in a document with read.table and prepares the input to be in the form
needed to later call \code{\link{createPage}} on it to create and display a
html page containing text and or various input format offered by shiny.
}
\details{
Input files must have a very specific form described in detail in the
vignettes (for details call ShinyPsych_Guide())
and can be read in from a local place or from dropbox. If dropbox is used
you must specify tokens to access it (see \code{\link[rdrop2]{drop_read_csv}})
in an rds file.
Available default files are "Demographics", "Goodbye",
"Goodbye_BanditFullStudy", "Instructions_Bandit",
"Instructions_BanditFullStudy", "Instructions_Bart", "Instructions_CheckId",
"Instructions_Dd", "Instructions_Dfd", "Instructions_Dfe",
"Instructions_Survey", "Instructions_TagsInput", "REI_Inventory",
"Survey_BanditFullStudy", "Survey_Example" and "TagsInput_Example".
}
\examples{
example.survey <- createPageList(fileName = "Survey_Example", randomize = TRUE)
example.survey
rm(example.survey)
}
