% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_task_ctrl_list.R
\name{createTaskCtrlList}
\alias{createTaskCtrlList}
\title{Create a List of Reactive Values for Task Control}
\usage{
createTaskCtrlList(task, oVar = NULL, oVarVals = NULL)
}
\arguments{
\item{task}{string. The task name. Has to be a single string since
for different tasks some variables are the same. If you have several tasks
in the same app just call this function several times. Valid inputs are
"bandit", "bart", "dfe", "dfd" and "dd".}

\item{oVar}{string. Vector with the name(s) of additional
variables that should be added to ctrlList.}

\item{oVarVals}{sting or numeric. Vector containing the values that should be
appended to the oVars variables. Must be the same length as oVars.}
}
\value{
A list of reactive values used for task control.
}
\description{
Create a list of reactive values ((see \code{\link[shiny]{reactiveValues}} or
\url{https://shiny.rstudio.com/reference/shiny/latest/reactiveValues.html}))
for task control. Additional variables can be specified.
}
\examples{
GameData <- createTaskCtrlList(task = "dfd")
print(shiny::isolate(shiny::reactiveValuesToList(GameData)))
rm(GameData)
}
