% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_page.R
\name{ddPage}
\alias{ddPage}
\title{Set up HTML Logic for the Delay Discounting Task}
\usage{
ddPage(ctrlVals, container, trialTitle = "Trial",
  trialTitlePractice = "Practice Trial", withPracticeTrial = FALSE,
  displayTitle = TRUE, nTrials = 10)
}
\arguments{
\item{ctrlVals}{list of reactive values. Controls the flow through the
Experiment. Can be created with \code{\link{createCtrlList}}.}

\item{container}{list. Containins the outcome time pairs and other
information on how to set up the tasks. Can be created with
\code{\link{createDdList}}}

\item{trialTitle}{string. Is displayed as title of the trial, e.g. if it is
set to "Trial" the displayed title in trial 2 of 10 will be "Trial 2 of 10".
Only needed if displayTitle is TRUE.}

\item{trialTitlePractice}{string. Is displayed as title of the practice
trial. Only needed if displayTitle is TRUE.}

\item{withPracticeTrial}{logical. If TRUE (default is FALSE) a practice
trial has to be provided and is included.}

\item{displayTitle}{logical. If TRUE (default) a title is displayed.}

\item{nTrials}{integer. The number of trials played WITHOUT the practice
trial.}
}
\value{
An html page that can be displayed by the shiny app.
}
\description{
Create html page from a list created by \code{\link{createDdList}} and
control client side communication.
}
