% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandit_custom_message.R
\name{.banditCustomMessage}
\alias{.banditCustomMessage}
\title{Send Custom Message for Bandit Task from R to Javascript in Browser}
\usage{
.banditCustomMessage(Arms, contrlVals, distribList, sess, containerOb,
  rDigits)
}
\arguments{
\item{Arms}{integer. The number of arms of the bandit. Has to be between 2
and 6.}

\item{contrlVals}{list of reactive values. Controls the flow through the
Experiment. Can be created with \code{\link{createCtrlList}}.}

\item{distribList}{list. Must at least contain the number of trials "nTrials"
(usually the one used to create the distributions with
\code{\link{createBanditList}}).}

\item{sess}{The session object from the shiny app.}

\item{containerOb}{list. Contains the outcomes of all bandit games.
Usually this list is created with \code{\link{createBanditList}}.}

\item{rDigits}{integer. Is passed to \code{\link[base]{round}}. Indicates
the number of digits the total point value displayed should be rounded to.}
}
\value{
An html page then displayed by the shiny app.
}
\description{
Send the information needed for the bandit task from R to javascript. What
exact information has to be sent depends on the number of arms specified.
More on custom messages on
\url{https://shiny.rstudio.com/reference/shiny/latest/session.html}
}
