% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfe_custom_message.R
\name{.dfeCustomMessage}
\alias{.dfeCustomMessage}
\title{Send Information from R Session to Browser}
\usage{
.dfeCustomMessage(Opts, contrlVals, sess, containerOb, signCols)
}
\arguments{
\item{Opts}{integer. The number of options to choose from.}

\item{contrlVals}{list of reactive values. Controls the flow through the
Experiment. Can be created with \code{\link{createCtrlList}}.}

\item{sess}{The session object from the shiny app.}

\item{containerOb}{list. Containing the outcomes. Can be created by calling
\code{\link{createDfePage}}.}

\item{signCols}{numeric. Controls the color of the outcome values
displayed. There are two options: 1 means negative values will be displayed
in red zeros in gray and positive values in green. A value other than 1 means
that all outcomes are displayed in black.}
}
\value{
Sends information to javascript where it is processed.
}
\description{
Sends outcome values for the decisions from experience task, usually created
with \code{\link{createDfePage}}, to the browser where they are then
processed and displayed by using java script.
}
