% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_values.R
\name{getValues}
\alias{getValues}
\title{Get Questionnaire Answers from Shiny Input and Append Them to a List}
\usage{
getValues(pageList, nmrc, inputList)
}
\arguments{
\item{pageList}{list. The page parameters, such as ids and text to
display. Create this with \code{\link{createPageList}}.}

\item{nmrc}{logical. If TRUE it is assumed that input was saved as numbers
and as.numeric will be called on the vector after processing. This
ensures that it's possible to e.g. calculate a mean of the vector later on.
If FALSE values will be returned as characters.}

\item{inputList}{The input object from the shiny app.}
}
\value{
A list containing the user inputs to the questionnaire.
}
\description{
Takes answers from the shiny input objects and appends them to a list. For
this to work, pageList must be created with \code{\link{createPageList}} or
have the same structure. If there are reversed scales in a questionnaire it
will also reverse them.
}
