% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/include_script_files.R
\name{includeScriptFiles}
\alias{includeScriptFiles}
\title{Load Javascript and CSS Files}
\usage{
includeScriptFiles(fileList = "", globalScript = TRUE, nArms = 0,
  nOpts = 0)
}
\arguments{
\item{fileList}{string. Vector with names to specify the tasks for which
the files should be loaded. Valid inputs are "bandit", "dfd", "dfe", "dd",
"bart" or any combination of these.}

\item{globalScript}{logical. If TRUE (default) the global style
script (not just used for the tasks) is loaded. Script does not need to be
named in fileList.}

\item{nArms}{integer. The number of arms the bandit should have. Only needed
if "bandit" is in fileList.}

\item{nOpts}{integer. The number of options in a trial are presented. Only
needed if fileList contains one of "dfe" or "dfd". If "dfe" the number must
be between 2 and 6, if "dfd" the number must be between 2 and 4.}
}
\value{
Loaded css and javascript scripts.
}
\description{
The different tasks for which ShinyPsych provides a framework often work with
javascript and are styled up a bit with css. This function loads the
appropriate scripts.
}
