% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_armed_bandit_page.R
\name{multiArmedBanditPage}
\alias{multiArmedBanditPage}
\title{Call Appropriate Create Bandit Page Function and Send Data to Browser}
\usage{
multiArmedBanditPage(ctrlVals, nArms, distList, container, roundDigits,
  session, ...)
}
\arguments{
\item{ctrlVals}{list of reactive values. Controls the flow through the
Experiment. Can be created with \code{\link{createCtrlList}}.}

\item{nArms}{integer. Number of arms the bandit has. Must be between 2 and 6}

\item{distList}{list. The one used to create the outcome lists with
\code{\link{createBanditList}}.}

\item{container}{list. Output of \code{\link{createBanditList}} containing the
outcome values to be displayed in the game.}

\item{roundDigits}{integer. The number of digits after the comma the total
point value should be rounded to.}

\item{session}{The session object from shiny}

\item{...}{Further arguments passed to the nArmedBanditPage functions
(e.g. \code{\link{twoArmedBanditPage}}).}
}
\value{
An html page that can be displayed by the shiny app. Also information
 is sent to javascript to animate the page.
}
\description{
Depending on the number of armes the bandit has, different create page
functions are needed and different information has to be sent to the browser.
This function takes care of this such that independend of the number of arms
only one function has to be called to set up the html and javascript logic.
}
