% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_opts_dfd_page.R
\name{multiOptsDfdPage}
\alias{multiOptsDfdPage}
\title{Call Appropriate Create Dfd Page Function}
\usage{
multiOptsDfdPage(ctrlVals, container, ...)
}
\arguments{
\item{ctrlVals}{list of reactive values. Controls the flow through the
Experiment. Can be created with \code{\link{createCtrlList}}.}

\item{container}{list. Output of \code{\link{createDfdList}} containing the
gambles to be displayed in a trial.}

\item{...}{Further arguments passed to the nOptDfdPage function called (e.g.
\code{\link{twoOptDfdPage}}).}
}
\value{
An html page that can be displayed by the shiny app.
}
\description{
Depending on the number of options a trial has, different create page
functions are needed. This function takes care of this such that independend
of the number of options only one function has to be called to set up the html
logic.
}
