% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_opts_dfe_page.R
\name{multiOptsDfePage}
\alias{multiOptsDfePage}
\title{Call Appropriate Create Dfe Page Function and Send Data to Browser}
\usage{
multiOptsDfePage(ctrlVals, session, container, signalColors, ...)
}
\arguments{
\item{ctrlVals}{list of reactive values. Controls the flow through the
Experiment. Can be created with \code{\link{createCtrlList}}.}

\item{session}{The session object of the shiny app.}

\item{container}{list. Output of \code{\link{createDfeList}} containing the
outcome values to be displayed in a trial.}

\item{signalColors}{numeric. Controls the color of the outcome values
displayed. There are two options: 1 means negative values will be displayed
in red zeros in gray and positive values in green. A value other than 1 means
that all outcomes are displayed in black.}

\item{...}{Further arguments passed to the nOptDfePage function called (e.g.
\code{\link{twoOptDfePage}}).}
}
\value{
An html page that can be displayed by the shiny app. Also information
 will be sent to javascript to animate the page.
}
\description{
Depending on the number of options the gamble has, different create page
functions are needed and different information has to be sent to the browser.
This function takes care of this such that independend of the number of options
only one function has to be called to set up the html and javascript logic.
}
