% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/next_bandit_page.R
\name{nextBanditPage}
\alias{nextBanditPage}
\title{Go to Next Page After a Game of the Bandit Task}
\usage{
nextBanditPage(ctrlVals, distList, gameData, withPracticeGame = TRUE,
  afterPracticePage = "postPractice", afterGamePage = "endGame",
  afterLastGamePage = "lastEndGame")
}
\arguments{
\item{ctrlVals}{list of reactive values. Controls the flow through the
Experiment. Can be created with \code{\link{createCtrlList}}.}

\item{distList}{list. The one used to create the outcome lists with
\code{\link{createBanditList}}.}

\item{gameData}{list of reactive values. Used for task control. Created with
\code{\link{createTaskCtrlList}}}

\item{withPracticeGame}{logical. If TRUE (default) the page after the first
game will be afterPracticePage, else it will be afterGamePage.}

\item{afterPracticePage}{string. The name of the page to be shown
after the practice trial. Only needs to be specified if withPracticeGamble
is set to TRUE.}

\item{afterGamePage}{string. The name of the page to be shown
after the trials up to the second to last trial.}

\item{afterLastGamePage}{string. The name of the page to be shown
after the last trial.}
}
\value{
An updated version of ctrlVals. Does not have to be assigned, since
 it is a reactive value.
}
\description{
If a game of the bandit task is finished this function will (if properly set
up, this has to be handled in the app) be called if the continue button is
pressed.
}
