% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/on_bart_action.R
\name{onBartAction}
\alias{onBartAction}
\title{Controls What Happens After a Bart Trial}
\usage{
onBartAction(id, ctrlVals, inputList, bartCtrlList, balloonList,
  nextPageId = NULL)
}
\arguments{
\item{id}{string. Controls what action should be executed (e.g. append values
or go to next page). Must be one of "nextballoon", "next_page", "popped" or
saveballoon.}

\item{ctrlVals}{list of reactive values. Controls the flow through the
Experiment. Can be created with \code{\link{createCtrlList}}.}

\item{inputList}{The input object from the shiny app.}

\item{bartCtrlList}{list of reactive values. Used for task control. Create
with \code{\link{createTaskCtrlList}}.}

\item{balloonList}{list. Output of \code{\link{createBartList}} containing the
game parameters.}

\item{nextPageId}{string. Id of next page that should be displayed. Only
needed if id is "next_page".}
}
\value{
Updated ctrlVals and bartCtrlList. Doesn't need to be assigned, since
 both are reactive values.
}
\description{
Controls which variables are stored in case of different events such as a
saved balloon or a popped balloon.
}
