% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/on_input_enable.R
\name{onInputEnable}
\alias{onInputEnable}
\title{Enable a Continue Button if a Specified Criterium is Reached}
\usage{
onInputEnable(pageId, ctrlVals, pageList, globId, inputList,
  charNum = NULL)
}
\arguments{
\item{pageId}{string. The current page id.}

\item{ctrlVals}{list of reactive values. Controls the flow through the
Experiment. Can be created with \code{\link{createCtrlList}}.}

\item{pageList}{list. The one returned by \code{\link{createPageList}} or one
of the same structure.}

\item{globId}{string. The global id of the set of pages or questionnaire as
specified in \code{\link{createCtrlList}}.}

\item{inputList}{The input object of the shiny app.}

\item{charNum}{integer. The minimum number of characters that must be given
to \code{\link[shiny]{textInput}} to enable the button. Only needed if
the page contains a text input.}
}
\value{
Updated state of the specified button.
}
\description{
Check if a criterium, specified in the page list file given or if default
loaded by \code{\link{createPageList}}, is met and if so enable the continue
button. If several criteria are given they must ALL be met to enable the
button.
}
