% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_data.R
\name{saveData}
\alias{saveData}
\title{Write Data to a Local Location or to Dropbox}
\usage{
saveData(data, location, partId, checkNull = TRUE, addNameList = NULL,
  suffix = "_s", outputDir = NULL, droptoken = "droptoken.rds",
  asrds = FALSE, separator = ",", mailSender = NULL,
  mailReceiver = NULL, mailSubject = "ShinyPsych Data",
  mailBody = "Data attached...")
}
\arguments{
\item{data}{list. Contains all the data that should be saved.}

\item{location}{string. The location to which data should be saved. Valid
inputs are "local" , "mail" or "dropbox".}

\item{partId}{string. The participant id. This will be pasted to the beginning
of the filename. If no id should be part of the filename leave this, in this
case a sample of length 9 from 1:9 will be drawn and used as first part of
the filename.}

\item{checkNull}{logical. If TRUE (default), each value in the list that is
NULL will be replaced by NA.}

\item{addNameList}{string. Vector with length of data containing the variable
names that should be given to the data. If names should not be change just
leave this as NULL.}

\item{suffix}{string. The suffix given to the name of the saved file. You
could e.g. use "_s" (default) for the survey data and "_g" for the game data
to have an easy distinction in the filenames.}

\item{outputDir}{string. Directory WITHOUT the file name in which to save the
data.}

\item{droptoken}{string. Name of the rds file with the dropbox acces tokens.
If this file is not in the app directory the full directory must be given.
Only needed if location is set to "dropbox".}

\item{asrds}{logical. If TRUE (default is FALSE), data is saved as an rds
file.}

\item{separator}{string. Separator that should be used to write the file.
Default is ",".}

\item{mailSender}{string. Mail sender address. Passed to
\code{\link[sendmailR]{sendmail}} from the sendmailR package.}

\item{mailReceiver}{string. Mail receiver address. Passed to
\code{\link[sendmailR]{sendmail}} from the sendmailR package.}

\item{mailSubject}{string. The mail subject. Passed to
\code{\link[sendmailR]{sendmail}} from the sendmailR package.}

\item{mailBody}{string. The mail body. Passed to
\code{\link[sendmailR]{sendmail}} from the sendmailR package.}
}
\value{
incProcess for shiny to display a process bar. Data is written out.
}
\description{
Take a list of a specified format and save it to the specified location. All
elements in the list must be either of same length or of length 1 because
\code{\link[base]{as.data.frame}} will be called on the list. The file name is
created by pasting partId, Sys.time(), digest(data) and the suffix (and of
course the file specification .rds or .csv).
}
