% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinypsych_guide.R
\name{ShinyPsych_Guide}
\alias{ShinyPsych_Guide}
\title{Open the ShinyPsych Guide}
\usage{
ShinyPsych_Guide()
}
\description{
Open the ShinyPsych global vignette that also contains link to the different
tutorial vignettes.
}
