% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_armed_bandit_page.R
\name{twoArmedBanditPage}
\alias{twoArmedBanditPage}
\alias{threeArmedBanditPage}
\alias{fourArmedBanditPage}
\alias{fiveArmedBanditPage}
\alias{sixArmedBanditPage}
\title{Create an HTML Page to Display and Play a 2 to 6 Armed Bandit}
\usage{
twoArmedBanditPage(ctrlValsList, gameTitle = "Game ", nTrials = NULL,
  clickCounter = "Clicks Remaining",
  gameTitlePractice = "Practice Game", withPracticeGame = TRUE,
  pointTitle = "Points Earned", nGames = 10,
  buttonLabel = "Click to Continue to next Game")

threeArmedBanditPage(ctrlValsList, gameTitle = "Game ", nTrials = NULL,
  clickCounter = "Clicks Remaining",
  gameTitlePractice = "Practice Game", withPracticeGame = TRUE,
  pointTitle = "Points Earned", nGames = 10,
  buttonLabel = "Click to Continue to next Game")

fourArmedBanditPage(ctrlValsList, gameTitle = "Game ", nTrials = NULL,
  clickCounter = "Clicks Remaining",
  gameTitlePractice = "Practice Game", withPracticeGame = TRUE,
  pointTitle = "Points Earned", nGames = 10,
  buttonLabel = "Click to Continue to next Game")

fiveArmedBanditPage(ctrlValsList, gameTitle = "Game ", nTrials = NULL,
  clickCounter = "Clicks Remaining",
  gameTitlePractice = "Practice Game", withPracticeGame = TRUE,
  pointTitle = "Points Earned", nGames = 10,
  buttonLabel = "Click to Continue to next Game")

sixArmedBanditPage(ctrlValsList, gameTitle = "Game ", nTrials = NULL,
  clickCounter = "Clicks Remaining",
  gameTitlePractice = "Practice Game", withPracticeGame = TRUE,
  pointTitle = "Points Earned", nGames = 10,
  buttonLabel = "Click to Continue to next Game")
}
\arguments{
\item{ctrlValsList}{list of reactive values. Controls the flow through the
Experiment. Can be created with \code{\link{createCtrlList}}.}

\item{gameTitle}{string. Is displayed as title of the trial, e.g. if it is
set to "Game" the displayed title in game 2 of 10 will be "Game 2 of 10".}

\item{nTrials}{integer. The number of games played WITHOUT the practice game.}

\item{clickCounter}{string. Is displayed next to the number of clicks still
left in one game. Default is set to "Clicks Remaining".}

\item{gameTitlePractice}{string. Is displayed as title of the practice game.}

\item{withPracticeGame}{logical. If TRUE (default) a practice
game must be provided and will be displayed.}

\item{pointTitle}{string. Is displayed as label next to the number of points
earned. Default is set to "Points Earned".}

\item{nGames}{integer. The number of games played WITHOUT the practice game.}

\item{buttonLabel}{string. Is displayed in the continue button. Default is
set to "Click to Continue to next Game".}
}
\value{
An html page that can be displayed by the shiny app.
}
\description{
Create an html page and send information to javascript if an option is
clicked. Needs prepared lists of outcome values created by
\code{\link{createBanditList}} or of the same structure and control variables
created with \code{\link{createCtrlList}}.
}
\details{
These functions are called from within \code{\link{multiArmedBanditPage}}
because this also handles the appropriate custom message.
}
\section{Functions}{
\itemize{
\item \code{threeArmedBanditPage}: A three armed bandit version

\item \code{fourArmedBanditPage}: A four armed bandit version

\item \code{fiveArmedBanditPage}: A five armed bandit version

\item \code{sixArmedBanditPage}: A six armed bandit version
}}

