% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_opt_dfd_page.R
\name{twoOptDfdPage}
\alias{twoOptDfdPage}
\alias{threeOptDfdPage}
\alias{fourOptDfdPage}
\title{Create an HTML Page to Display and Play a Decisions From Description Task}
\usage{
twoOptDfdPage(ctrlValsList, containerOb, gambleTitle = "Gamble",
  gambleTitlePractice = "Practice Gamble", withPracticeGamble = TRUE,
  displayTitle = TRUE, nGambles = 10)

threeOptDfdPage(ctrlValsList, containerOb, gambleTitle = "Gamble",
  gambleTitlePractice = "Practice Gamble", withPracticeGamble = TRUE,
  displayTitle = TRUE, nGambles = 10)

fourOptDfdPage(ctrlValsList, containerOb, gambleTitle = "Gamble",
  gambleTitlePractice = "Practice Gamble", withPracticeGamble = TRUE,
  displayTitle = TRUE, nGambles = 10)
}
\arguments{
\item{ctrlValsList}{list of reactive values. Controls the flow through the
Experiment. Can be created with \code{\link{createCtrlList}}.}

\item{containerOb}{list. Contains the gambles. Can be created by calling
\code{\link{multiOptsDfdPage}}.}

\item{gambleTitle}{string. Is displayed as title of the trial, e.g. if it is
set to "Gamble" the displayed title in game 2 of 10  will be "Gamble 2 of
10".}

\item{gambleTitlePractice}{string. Is displayed as title of the practice
gamble.}

\item{withPracticeGamble}{logical. If TRUE (default) a practice
gamble must be provided and is played. Only needed if displayTitle is TRUE.}

\item{displayTitle}{logical. If TRUE (default) a title is displayed.}

\item{nGambles}{integer. The number of gambles played WITHOUT the practice
game.}
}
\value{
An html page that can be displayed by a shiny app.
}
\description{
Create an html page and send information to javascript if an option is
clicked. Needs prepared lists of gambles created by
\code{\link{createDfdList}} or of the same structure and control variables
created with \code{\link{createCtrlList}}.
}
\details{
These functions are called from within \code{\link{multiOptsDfdPage}}
because this also handles the appropriate custom message. So when you need
this function just call \code{\link{multiOptsDfdPage}}.
}
\section{Functions}{
\itemize{
\item \code{threeOptDfdPage}: A three option gamble version

\item \code{fourOptDfdPage}: A four option gamble version
}}

