% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_opt_dfe_page.R
\name{twoOptDfePage}
\alias{twoOptDfePage}
\alias{threeOptDfePage}
\alias{fourOptDfePage}
\alias{fiveOptDfePage}
\alias{sixOptDfePage}
\title{Create an HTML Page to Display and Play a Decisions From Experience Task}
\usage{
twoOptDfePage(ctrlValsList, gambleTitle = "Gamble",
  gambleTitlePractice = "Practice Gamble", withPracticeGamble = TRUE,
  displayTitle = TRUE, nGambles = 10, buttonLabel = "Make Decision")

threeOptDfePage(ctrlValsList, gambleTitle = "Gamble",
  gambleTitlePractice = "Practice Gamble", withPracticeGamble = TRUE,
  displayTitle = TRUE, nGambles = 10, buttonLabel = "Make Decision")

fourOptDfePage(ctrlValsList, gambleTitle = "Gamble",
  gambleTitlePractice = "Practice Gamble", withPracticeGamble = TRUE,
  displayTitle = TRUE, nGambles = 10, buttonLabel = "Make Decision")

fiveOptDfePage(ctrlValsList, gambleTitle = "Gamble",
  gambleTitlePractice = "Practice Gamble", withPracticeGamble = TRUE,
  displayTitle = TRUE, nGambles = 10, buttonLabel = "Make Decision")

sixOptDfePage(ctrlValsList, gambleTitle = "Gamble",
  gambleTitlePractice = "Practice Gamble", withPracticeGamble = TRUE,
  displayTitle = TRUE, nGambles = 10, buttonLabel = "Make Decision")
}
\arguments{
\item{ctrlValsList}{list of reactive values. Controls the flow through the
Experiment. Can be created with \code{\link{createCtrlList}}.}

\item{gambleTitle}{string. Is displayed as title of the trial, e.g. if it is
set to "Gamble" the displayed title in game 2 of 10  will be "Gamble 2 of
10".}

\item{gambleTitlePractice}{string. Is displayed as title of the practice
gamble.}

\item{withPracticeGamble}{logical. If TRUE (default) a practice
gamble must be provided and is played. Only needed if displayTitle is TRUE.}

\item{displayTitle}{logical. If TRUE (default) a title is displayed.}

\item{nGambles}{integer. The number of gambles played WITHOUT the practice
game.}

\item{buttonLabel}{string. Is displayed in the continue button. Default is
set to "Make Decision".}
}
\value{
An html page that can be displayed by a shiny appl
}
\description{
Create an html page and send information to javascript if an option is
clicked. Needs prepared lists of outcomes created by
\code{\link{createDfeList}} or of the same structure and control variables
created with \code{\link{createCtrlList}}.
}
\details{
These functions are called from within \code{\link{multiOptsDfePage}}
because this also handles the appropriate custom message with.
}
\section{Functions}{
\itemize{
\item \code{threeOptDfePage}: A three option gamble version

\item \code{fourOptDfePage}: A four option gamble version

\item \code{fiveOptDfePage}: A five option gamble version

\item \code{sixOptDfePage}: A six option gamble version
}}

